rename use-component-filter to component-filter-context
